% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_Sj.R
\name{base_Sj}
\alias{base_Sj}
\title{Makes a table with certain sequencies in sample and their frequencies.}
\usage{
base_Sj(S, j, base_Sja, lenX, d)
}
\arguments{
\item{S}{A set of lags in \eqn{1:d}.}

\item{j}{A lag \eqn{j \in (1:d) \setminus S}.}

\item{base_Sja}{A table generated by base_Sja.}

\item{lenX}{The length of sample vector.}

\item{d}{The chains order.}
}
\value{
A table with sequencies x indexed by \eqn{S\cup j}, their frequencies and the MLE estimator of the transition probabilities.
}
\description{
Makes a table with certain sequencies in sample and their frequencies.
}
