% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_Sja.R
\name{base_Sja}
\alias{base_Sja}
\title{Makes a table with certain sequencies in sample and their frequencies.}
\usage{
base_Sja(S, j = NULL, A, base)
}
\arguments{
\item{S}{A set of lags in \eqn{1:d}.}

\item{j}{A lag \eqn{j \in (1:d) \setminus S}.}

\item{A}{The states space.}

\item{base}{A table with sequencies of size d+1 in sample and their absolute frequency.}
}
\value{
A table with sequencies (x) indexed by \eqn{S\cup j} followed by any element a, their frequencies and the MLE estimator of the transition probabilities.
}
\description{
Makes a table with certain sequencies in sample and their frequencies.
}
