% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC_S.R
\name{BIC_S}
\alias{BIC_S}
\title{Calculates de BIC given a set of relevant lags.}
\usage{
BIC_S(X, A, S = 1:d, xi = 1/2)
}
\arguments{
\item{X}{A Markov chain.}

\item{A}{The states space.}

\item{S}{A set of relevant lags, if empty S=\eqn{1,2,\dots, d}.}

\item{xi}{The BIC constant. Defaulted to 1/2. Smaller c \verb{(near 0)} reduces the impact of overparameterization.}
}
\value{
A vector of estimated BICs, the first entry supposes the first element
of S is the only relevant lag. The second entry supposes the first and the second
elements of S are the only relevant lags, and so on, until the penultimate entry
which supposes S is the relevant lag set. The last entry shows which of the previous
entries has the smallest BIC.
}
\description{
Given a sample, this function estimates the BIC of a string whose set of relevant
lags is just the first element of S. Then it assumes that the set of relevant lags
are the first 2 elements of S and also estimates the BIC, and so on...
Lastly, estimate the BIC assuming that the set of relevant lags is exactly the set S.
}
\details{
This function assumes that the order of elements in S is determined by the relevance
(i.e. the first element of S is the most relevant lag, and probably the
one with the greatest oscillation, the last element of S is the least relevant,
probably has the smallest oscillation).
}
