% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqTabSj.R
\name{freqTabSj}
\alias{freqTabSj}
\title{A tibble showing the frequency of sample sequences}
\usage{
freqTabSj(S, j, freqTab, lenX, d)
}
\arguments{
\item{S}{A numeric vector of positive integers or \code{NULL}. Represents a set
of past lags that must be present within the columns of the \code{freqTab} argument. Typically,
\code{S} is a subset of \code{1:d}.}

\item{j}{An integer or \code{NULL}. Typically represents a lag in the
complement of \code{S} (in \code{1:d}). Both \code{S} and \code{j} cannot be \code{NULL}
at the same time.}

\item{freqTab}{A tibble generated by the \code{\link[=freqTab]{freqTab()}} function, must contain columns referent to lags
\code{S} and \code{j}.}

\item{lenX}{A positive integer representing the number of elements in the sample.}

\item{d}{A positive integer representing an upper bound for the chain order.}
}
\value{
A tibble with sequences indexed by the lags in \code{S} and \code{j}, their frequencies,
and the maximum likelihood estimator (MLE) estimator of the transition probabilities.
}
\description{
Creates a tibble that displays the frequencies of sequences within the sample.
}
