% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_xS.R
\name{is_xS}
\alias{is_xS}
\title{Checks which row in data frame is equal to a vector}
\usage{
is_xS(x, y)
}
\arguments{
\item{x}{A data frame whose rows are sequencies indexed by S.}

\item{y}{A specific x_S sequence.}
}
\value{
A vector of true and false if a row in x is exactly equal to x_S.
}
\description{
Checks which row in data frame is equal to a vector
}
