% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_parameters.R
\name{n_parameters}
\alias{n_parameters}
\title{Number of parameters of a MTD model.}
\usage{
n_parameters(Lambda, A, single_matrix = FALSE, indep_part = TRUE)
}
\arguments{
\item{Lambda}{The relevant lag set.}

\item{A}{The states space.}

\item{single_matrix}{If TRUE all p_j matrix are equal.}

\item{indep_part}{If FALSE independent distribution is set to zero.}
}
\value{
The number of parameters of a MTD model.
}
\description{
Calculates the number of parameters of a MTD model with relevant lag set Lambda and states space A.
}
