% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oscillationMax.R
\name{oscillationMax}
\alias{oscillationMax}
\title{Oscillations of a MTD Markov chain}
\usage{
oscillationMax(x, ...)
}
\arguments{
\item{x}{Must be a MTD object or a chain sample.}

\item{...}{bla}
}
\value{
If the x parameter is an MTD object, it will provide the oscillations for
each relevant element. In case x is a MTD chain sample, it estimates the oscillations
for a user-inputted set S of lags.
}
\description{
Calculates the oscillations of a MTD model object or estimates the oscillations of a chain sample.
}
\details{
The oscillations of a MTD model
(\eqn{\delta_k} for \eqn{k in \Lambda}), are the product of the weight \eqn{\lambda_k}
multiplied by the maximum of the total variation distances between the
distributions in the matrix p_k. These values are important because they
measure the influence of a relevant lag k on the model.
}
\examples{
oscillationMax( MTDmodel(Lambda=c(1,4),A=c(2,3) ) )
oscillationMax( MTDmodel(Lambda=c(1,4),A=c(2,3),w0=0.01 ) )
oscillationMax( MTDmodel(Lambda=c(1,4),A=c(2,3),w0=0.01,w_j=c(0.49,0.5)) )
oscillationMax( MTDmodel(Lambda=c(1,4),A=c(2,3),w0=0.01,w_j=c(0.49,0.5),
p_j=list(matrix(c(0.1,0.9,0.9,0.1),ncol=2)), single_matrix=TRUE))

}
