% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseMarkov.R
\name{sparseMarkov}
\alias{sparseMarkov}
\title{A function for inference in sparse Markov chains}
\usage{
sparseMarkov(X, d, method, l = NULL, alpha = 0.05, mu = 1, xi = 0.5)
}
\arguments{
\item{X}{A markov chain sample}

\item{d}{The order or a upper bound for the order.}

\item{method}{A method for estimation}

\item{l}{A stop point for some methods}

\item{alpha}{CUT parameter}

\item{mu}{CUT parameter}

\item{xi}{CUT parameter}
}
\value{
The relevant lag set
}
\description{
A function for inference in sparse Markov chains
}
