% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseMarkov_FS.R
\name{sparseMarkov_FS}
\alias{sparseMarkov_FS}
\title{Applies Forward Stepwise algorithm}
\usage{
sparseMarkov_FS(X, A, d, l)
}
\arguments{
\item{X}{A mixture transition distribution chain sample.}

\item{A}{The States space.}

\item{d}{The chains order.}

\item{l}{Stop point for FS algorithm.}
}
\value{
Returns a set S of size l that contains relevant lags.
}
\description{
Applies Forward Stepwise algorithm
}
