% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseMarkov_FSC.R
\name{sparseMarkov_FSC}
\alias{sparseMarkov_FSC}
\title{A function for inference in MTD Markov chains}
\usage{
sparseMarkov_FSC(
  X,
  A = NULL,
  d,
  l = NULL,
  alpha = 0.05,
  mu = 1,
  xi = 0.5,
  warning = FALSE
)
}
\arguments{
\item{X}{A MTD Markov chain sample.}

\item{A}{The states space.}

\item{d}{The chains order.}

\item{l}{Stop point for FS algorithm.}

\item{alpha}{A parameter of CUT.}

\item{mu}{A parameter of CUT.}

\item{xi}{A parameter of CUT.}

\item{warning}{If TRUE may return warnings.}
}
\value{
Returns a estimated set of relevant lags.
}
\description{
This function estimates the relevant lag set \eqn{\Lambda}
of a MTD model through the FSC algorithm.
}
\details{
The "Forward Stepwise and Cut" (FSC)is an algorithm for inference in
Mixture Transition Ditribution (MTD) models.
It consists in the application of the "Forward Stepwise" (FS) step followed by the CUT algorithm.
This method was developed by \href{https://arxiv.org/abs/2202.08007}{Ost and Takahashi} and is specially useful for High order MTD Markov chains.
}
