% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{test2_chain}
\alias{test2_chain}
\title{A chain for testing the package}
\format{
A data frame with 10000 rows and 2 variables:
\describe{
\item{X}{d=15, L=1,5,10, MTD: .3,.3,.3}
\item{Y}{d=30, L=1,10,18, MTD: .15,.35,.4}
}
}
\source{
{Created in-house to serve as example.}
}
\usage{
test2_chain
}
\description{
A data frame with Markov Chain samples with Misture Transition
distribution.
}
\examples{
data(test2_chain)
}
\keyword{datasets}
