% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_chain}
\alias{test_chain}
\title{A chain for testing the package}
\format{
A data frame with 10000 rows and 1 variables:
\describe{
\item{Xn}{a Markov Chain sample}
}
}
\source{
{Created in-house to serve as example.}
}
\usage{
test_chain
}
\description{
A single column data frame with a Markov Chain sample with Misture Transition
distribution.
}
\examples{
data(test_chain)
}
\keyword{datasets}
