% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_Sja.R
\name{base_Sja}
\alias{base_Sja}
\title{Makes a table with certain sequences in sample and their frequencies.}
\usage{
base_Sja(S, j = NULL, A, base, complete = TRUE)
}
\arguments{
\item{S}{A set of lags in \eqn{1:d}.}

\item{j}{A lag \eqn{j \in (1:d) \setminus S}.}

\item{A}{The states space.}

\item{base}{A table with sequences of size d+1 in sample and their absolute frequency.}

\item{complete}{logical, if TRUE and there are sequences that didn't appear in the sample,
the base will be completed with such sequences appearing 0 times.}
}
\value{
A table with sequences x indexed by \eqn{S\cup j} followed by any element a, their frequencies and the MLE estimator of the transition probabilities.
}
\description{
Makes a table with certain sequences in sample and their frequencies.
}
