% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC_l.R
\name{BIC_l}
\alias{BIC_l}
\title{Bayesian Information Criterion \code{(BIC)} of a Markov Chain.}
\usage{
BIC_l(X, d, l, c = 1/2)
}
\arguments{
\item{X}{Markov chain.}

\item{d}{A upper threshold for the chains order.}

\item{l}{A upper threshold for the number of elements in the relevant lag set.}

\item{c}{The BIC constant. Defaulted to 1/2. Smaller c \verb{(near 0)} reduces the impact of overparameterization.}
}
\value{
The BIC of a Markov Chain for each possible set of lags with size 1,2...,l.
}
\description{
Bayesian Information Criterion \code{(BIC)} of a Markov Chain.
}
