% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMTD.R
\name{checkMTD}
\alias{checkMTD}
\title{Checks a MTD object}
\usage{
checkMTD(MTD)
}
\arguments{
\item{MTD}{A MTD object.}
}
\description{
Verifies if a MTD object is correctly structured, containing all the necessary
parameters and if these parameters satisfy their respective constraints.
}
