% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMTD.R
\name{checkMTD}
\alias{checkMTD}
\title{Checks a class MTD object}
\usage{
checkMTD(MTD)
}
\arguments{
\item{MTD}{An object of class MTD. Must be properly structured as such. See \emph{Details} below for
more information.}
}
\description{
Verifies if an object of class MTD is correctly structured, containing all the necessary
parameters, and if these parameters satisfy their respective constraints.
}
\details{
A properly defined class MTD object can be created through the \code{\link[=MTDmodel]{MTDmodel()}}.
To verify the correct structure and instances of a class MTD object, please check \code{\link[=MTDmodel]{MTDmodel()}}
documentation.
}
