% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSample.R
\name{checkSample}
\alias{checkSample}
\title{Checks if a sample is suitable for use.}
\usage{
checkSample(X)
}
\arguments{
\item{X}{A vector with a Markov Chain.}
}
\value{
Returns possible sample errors.
}
\description{
Checks if a sample is suitable for use.
}
\examples{
checkSample(c(1,1,2,1,1,2,1,2,1,1,2,2,1))
}
