% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSample.R
\name{checkSample}
\alias{checkSample}
\title{Check a sample}
\usage{
checkSample(X)
}
\arguments{
\item{X}{A vector with a sample of an MTD Markov chain.}
}
\value{
Returns an adjusted chain sample and possible sample errors.
}
\description{
Checks if a sample is suitable for use.
}
