% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSample.R
\name{checkSample}
\alias{checkSample}
\title{Checks a sample}
\usage{
checkSample(X)
}
\arguments{
\item{X}{A vector or a single-column data frame with a sample of an MTD chain.}
}
\value{
Returns the sample as a vector or identifies any possible sample problems.
}
\description{
Checks if a sample is a suitable argument for some functions within
the package.
}
