% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSample.R
\name{checkSample}
\alias{checkSample}
\title{Checks if a Markov chain sample is suitable for use}
\usage{
checkSample(X)
}
\arguments{
\item{X}{A vector with a Markov Chain.}
}
\value{
Returns adjusted sample and possible sample errors.
}
\description{
Checks if a Markov chain sample is suitable for use
}
