% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsTab.R
\name{countsTab}
\alias{countsTab}
\title{Table of sample sequence counts.}
\usage{
countsTab(X, d)
}
\arguments{
\item{X}{A sample from an MTD Markov Chain.}

\item{d}{An upper bound for the chain's order.}
}
\value{
A table with every size d+1 sequence in the sample and its absolute frequency.
}
\description{
Creates a table with sample sequences and their absolute frequencies.
}
\details{
The function will make a tibble with \eqn{d+2} columns. For each row, the first \eqn{d+1} columns will
have a sequence of size \eqn{d+1} that appeared in the sample. The last column, called \code{Nxa},
will contain the number of times each of these sequences appeared in the sample.
}
\examples{
countsTab(c(1,2,2,1,2,1,1,2,1,2),3)
countsTab(c(0,2,0,2,0,2,1,1,0,0,1,2,1,2,1),4)
}
