% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dTV_pj.R
\name{dTV_pj}
\alias{dTV_pj}
\title{Total variation distance for p_j distribution.}
\usage{
dTV_pj(p_j, rows)
}
\arguments{
\item{p_j}{A stochastic matrix.}

\item{rows}{A matrix with two columns, the pair of numbers in each row represents the rows of p_j to be compared,
i.e. the distributions in p_j to be compared.}
}
\value{
Returns the total variation distance of a set of distributions in p_j.
}
\description{
Calculates the total variation distance given a stochastic matrix p_j.
}
