% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dTV_pj.R
\name{dTV_pj}
\alias{dTV_pj}
\title{The maximum total variation distance}
\usage{
dTV_pj(pj, rows, which_max = FALSE)
}
\arguments{
\item{pj}{A stochastic matrix.}

\item{rows}{A numeric matrix with two columns. Each row must contain a pair of
numbers representing the indices of the rows in \code{pj} that will be compared,
i.e., the distributions.}

\item{which_max}{Logical. If TRUE, the function will also return the pair in
\code{rows} that represents the row indices of the matrix \code{pj} with the
greatest total variation distance.}
}
\value{
Returns the maximum total variation distance of a set of distributions in
\code{pj}. If \code{which_max=TRUE}, it returns a list containing the maximum total
variation distance along with the pair of row indices in \code{pj} representing the
distributions with the greatest distance.
}
\description{
Computes the total variation distance between any two rows of a stochastic matrix.
Can analyze several pairs of rows at once, and if so, will return the greatest
total variation distance calculated among the pairs.
}
