% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dTV_sample.R
\name{dTV_sample}
\alias{dTV_sample}
\title{The total variation distance between distributions}
\usage{
dTV_sample(S, j, A = NULL, base, lenA = NULL, A_pairs = NULL, x_S)
}
\arguments{
\item{S}{A numeric vector of positive integers or \code{NULL}. Represents
a set of past lags. The distributions from which this function will calculate the total
variation distance are conditioned on a fixed sequence indexed by \code{S} ( the user must
also input the sequence through the argument \code{x_S}).}

\item{j}{An integer number. Represents a lag \code{j} in the \eqn{complement} of \code{S}.
The symbols indexed by \code{j} vary along the state space \code{A}, altering the distribution
through this single lag, and the size of this change is what this function seeks to measure.}

\item{A}{A vector with positive integers. Must have at least two different entries. \code{A}
represents the state space. You may leave \code{A=NULL} (default) if you provide the function
with the arguments \code{lenA} and \code{A_pairs} (see description below).}

\item{base}{A data frame where each rows first \code{length(S)+2} columns represent a length
\code{length(S)+2} sequence of elements from \code{A}. The first column symbol represents the
one further in the past and so on until the symbol in column \code{length(S)+2} is the latest
in the sequence. Also, the data frame must have a column with the transition probabilities
from the sequence in the first \code{length(S)+1} columns to the symbol at column \code{length(S)+2}.
In other words, each row depicts a past sequence, a present symbol, and the probability of seeing
this symbol after that sequence occurred at times \code{S} and \code{j}. Given a \code{x_S}
past sequence, \code{base} must have at least every transition probability conditioned in
past \code{x_S}. This means that there must be \code{length(A)^2} rows with sequence \code{x_S},
(since the element in column \code{S+2} and the element in column 'xj' must vary across all
elements of \code{A}). Since this function was created to estimate total variation distances
between distributions calculated through the \code{\link[=freqTab]{freqTab()}} function, \code{base} is typically
an output from \code{\link[=freqTab]{freqTab()}} (the \code{base} format must match that of an output from \code{\link[=freqTab]{freqTab()}}
with the argument \code{complete=TRUE}). Please check the \code{\link[=freqTab]{freqTab()}} documentation and the
\emph{Details} section for further information.}

\item{lenA}{An integer greater than 1 representing the length of the state space A, i.e.,
\code{lenA <- length(A)}. If \code{A} is provided, there is no need to input \code{lenA}.}

\item{A_pairs}{A matrix with all possible pairs of elements from \code{A}. If \code{A} is provided,
there is no need to input \code{A_pairs}.}

\item{x_S}{A vector of length \code{length(S)} or \code{NULL}. If \code{S=NULL}, \code{x_S} will
be set to \code{NULL}. \code{x_S} represents a sequence of symbols from \code{A} indexed by
\code{S}. This sequence remains constant across the conditional distributions to be compared,
representing the fixed configuration of the past.}
}
\value{
Returns a vector of total variation distances, where each entry corresponds to the
distance between a pair of distributions conditioned on the same fixed past \code{x_S},
differing only in the symbol indexed by \code{j}, which varies across all distinct pairs
of elements in \code{A}.Therefore, the vector contains as many entries as there are
distinct pairs with elements of \code{A}.
}
\description{
Calculates the total variation distance between distributions conditioned
in a given past.
}
\details{
About the \code{base} argument format: this function was created to calculate
the total variation distance between distributions estimated through the \code{\link[=freqTab]{freqTab()}} function.
Hence, typically, \code{base} is an output from \code{\link[=freqTab]{freqTab()}}, and its format is imperative
for this function to work. If you wish to use \code{\link[=dTV_sample]{dTV_sample()}} to estimate total variation
distances for distributions acquired otherwise, please note that they must be inputted through
the \code{base} argument, following the same pattern as an output from \code{\link[=freqTab]{freqTab()}} (i.e.,
column names must match, the actual probabilities should be named 'qax_Sj', and the past
and present lags should be separated into different columns with appropriate column names.
However, the columns 'Nx_Sj' and 'Nxa_Sj' do not need to be present in \code{base}).

When you provide the state space \code{A} to this function, it will use it to calculate
the arguments \code{lenA <- length(A)} and \code{A_pairs <- t(utils::combn(A, 2))}. While
\code{A} is typically used, this function also allows for the input of \code{lenA} and
\code{A_pairs} instead of \code{A}. This flexibility is particularly useful within the
\code{hdMTD} package functions where \code{\link[=dTV_sample]{dTV_sample()}} may be repeatedly used in loops with
the same \code{A}, optimizing computational efficiency.
}
\examples{
#creating base argument through freqTab function.
pbase <- freqTab(S=c(1,4),j=2,A=c(1,2,3),countsTab = countsTab(testChains[,2],d=5))
dTV_sample(S=c(1,2),j=4,A=c(1,2,3),base=pbase,x_S=c(2,3))
pbase <- freqTab(S=NULL,j=1,A=c(1,2,3),countsTab = countsTab(testChains[,2],d=5))
dTV_sample(S=NULL,j=1,A=c(1,2,3),base=pbase)
}
