% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dTV_sample.R
\name{dTV_sample}
\alias{dTV_sample}
\title{Total variation distance between distributions given x_S}
\usage{
dTV_sample(S, j, lenA, base, A_pairs, x_S)
}
\arguments{
\item{S}{A set of lags.}

\item{j}{A specific lag \eqn{j in the complement of S}.}

\item{lenA}{The length of the states space A.}

\item{base}{A set with the estimated transition probabilities qa|xb_Sj.}

\item{A_pairs}{A list with all pairs with elements of A.}

\item{x_S}{A specific sequence indexed by S.}
}
\value{
Returns a vector with the total variation distance between estimated distributions
given a x_S (i.e. a past indexed by S).
}
\description{
Total variation distance between distributions given x_S
}
