% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dTV_sample.R
\name{dTV_sample}
\alias{dTV_sample}
\title{Total variation distance between distributions}
\usage{
dTV_sample(S, j, lenA, base, A_pairs, x_S)
}
\arguments{
\item{S}{A set of lags.}

\item{j}{A lag j in the \eqn{complement} of S.}

\item{lenA}{The length of the states space A.}

\item{base}{A set with the estimated transition probabilities \eqn{q(a|xb_Sj)}.}

\item{A_pairs}{A list with all pairs with elements of A.}

\item{x_S}{A specific sequence indexed by S.}
}
\value{
Returns a vector with the total variation distance between the estimated distributions
given a past x_S (i.e. a past x indexed by S).
}
\description{
Calculates the total variation distance between distributions conditioned in a given past x_S.
}
