% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqTab.R
\name{freqTab}
\alias{freqTab}
\title{Table of probabilities and frequencies of sample sequences}
\usage{
freqTab(S, j = NULL, A, countsTab, complete = TRUE)
}
\arguments{
\item{S}{A subset of \eqn{1:d}.}

\item{j}{An element j in set \eqn{ {1:d} \ {S} }.}

\item{A}{The states space.}

\item{countsTab}{A data frame with all sequences of size d+1 that appear in the sample, and their absolute frequency.
Such table is generated by the function \code{\link[=countsTab]{countsTab()}}. If any data frames that were not generated by \code{\link[=countsTab]{countsTab()}}
are used, problems may occur since the column names are important for \code{\link[=freqTab]{freqTab()}} function}

\item{complete}{Logical. If TRUE and if there are sequences that didn't appear in the sample,
the freqTab table will be completed with such sequences appearing 0 times.}
}
\value{
A table with the sequences that appeared in the sample (at times \eqn{ {-S}U{-j}U{0} } for each size d+1 sequence),
their frequencies and the MLE estimator of the transition probabilities.
}
\description{
A table with the sample sequences (including time 0) frequencies , and the estimated transition probabilities.
}
\examples{
freqTab(S=c(1,3),j=NULL,A=c(0,1,2),
countsTab=countsTab(c(1,0,1,0,2,0,1,0,1,0,0,0,0,1,0,1,0,0,0,0),d=3))
freqTab(S=c(1,3),j=NULL,A=c(0,1,2),
countsTab=countsTab(c(1,0,1,0,2,0,1,0,1,0,0,0,0,1,0,1,0,0,0,0),d=3),complete=FALSE)
}
