% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqTab.R
\name{freqTab}
\alias{freqTab}
\title{A tibble with probabilities and frequencies of sample sequences}
\usage{
freqTab(S, j = NULL, A, countsTab, complete = TRUE)
}
\arguments{
\item{S}{A numeric vector of positive integers or \code{NULL}. Represents a set
of past lags that must be present within the columns of the \code{countsTab} argument and are
to be considered while estimating the transition probabilities. Both \code{S} and \code{j}
cannot be \code{NULL} at the same time.}

\item{j}{An integer or \code{NULL}. Typically represents a lag \code{j} in the
\eqn{complement} of \code{S}. Both \code{S} and \code{j} cannot be \code{NULL} at the same time.
See \emph{Details} for further information.}

\item{A}{A vector with positive integers. Must have at least two different entries. \code{A}
represents the state space.}

\item{countsTab}{A tibble or a data frame with all sequences of length d+1 that appear in the sample,
and their absolute frequency. Such tibble typically is generated by the function \code{\link[=countsTab]{countsTab()}}.
If any data frames that were not generated by \code{\link[=countsTab]{countsTab()}} are used, problems may occur since
its format and column names are important within \code{\link[=freqTab]{freqTab()}} function.}

\item{complete}{Logical. If \code{TRUE} all sequences that didn't appear in the sample
will be included in the output with frequency equal to 0 .}
}
\value{
A tibble where each row represents a sequence of elements from \code{A}.
The initial columns display each sequence symbol separated into columns corresponding to their time indexes.
The remaining columns show the sample frequencies of the sequences and the MLE (Maximum Likelihood Estimator)
of the transition probabilities.
}
\description{
A tibble with the sample sequences, their frequencies and the estimated
transition probabilities.
}
\details{
Within this function, both arguments \code{S} and \code{j} perform the same role of
filtering the columns of the \code{countsTab} argument to be kept in the \code{freqTab()} output.
In fact, any specific lag \code{j} could be provided to this function via the \code{S} argument,
while leaving the \code{j} argument as \code{NULL} as defaulted. The output would be exactly the
same as when specifying \code{S} and \code{j} separately. The inclusion of \code{j} as a
parameter is merely for organizational purposes and is a useful tool for clarity within
the package's algorithms.
}
\examples{
freqTab(S=c(1,4),j=2,A=c(1,2,3),countsTab = countsTab(testChains[,2],d=5))
#Equivalent to freqTab(S=c(1,2,4),j=NULL,A=c(1,2,3),countsTab = countsTab(testChains[,2],d=5))
}
