% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqTabSj.R
\name{freqTabSj}
\alias{freqTabSj}
\title{A table with the frequency of sample sequences}
\usage{
freqTabSj(S, j, freqTab, lenX, d)
}
\arguments{
\item{S}{A subset of \eqn{1:d}.}

\item{j}{An element \eqn{j \in (1:d) \setminus S}.}

\item{freqTab}{A table generated by the base_Sja function.}

\item{lenX}{The length of sample vector.}

\item{d}{The chains order.}
}
\value{
A table with sequences x indexed by \eqn{{-S} U {-j}}, their frequencies and the MLE estimator of the transition probabilities.
}
\description{
Creates a table with the frequencies of the sequences in the sample.
}
