% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdMTD.R
\name{hdMTD}
\alias{hdMTD}
\title{Inference in MTD models}
\usage{
hdMTD(X, d, method = "FS", ...)
}
\arguments{
\item{X}{A Mixture Transition Distribution (MTD) chain sample.}

\item{d}{An upper bound for the chains order.}

\item{method}{A method for estimation of the relevant lag set. The methods available in this package are "FS" (default),"FSC","CUT" and "BIC". See documentation for
each method with its respective function as explained in "details".}

\item{...}{Extra arguments relevant to the chosen method, see "details" for more information.}
}
\value{
Returns the estimated relevant lag set for a MTD chain sample.
}
\description{
A function for inference in Mixture Transition Distribution (MTD) Markov chains. This function can use a selected "method" to perform estimation of the relevant lag set of a MTD chain sample.
The default method is "FS" (Foward Stepwise) which is specially useful in high dimension. The other available methods are "CUT", FSC" (Foward Stepwise and Cut) which is an application of the
"FS" method followed by the "CUT" method, and lastly the "BIC" (Bayesian Information Criterion) method. For more information on these methods see the documentation of their specific functions
listed on "details" below.
}
\details{
This function is simply a way to gather all of the hdMTD_ functions in a single place.
For example, if the \code{\link[=hdMTD]{hdMTD()}} function is used with method="FSC" it will call the \code{\link[=hdMTD_FSC]{hdMTD_FSC()}} function. Note that, in this case, any extra parameters must match those used by \code{\link[=hdMTD_FSC]{hdMTD_FSC()}}.
Each method may use a different set of parameters, and they can be passed to \code{\link[=hdMTD]{hdMTD()}} trough the ... argument. In other to see which parameters can be passed for each method
seek the documentation of the hdMTD_"method" function as follows:
\itemize{
\item For "FS" method, extra parameters are listed in \code{\link[=hdMTD_FS]{hdMTD_FS()}} documentation.
\item For "FSC" method, extra parameters are listed in \code{\link[=hdMTD_FSC]{hdMTD_FSC()}} documentation.
\item For "CUT" method, extra parameters are listed in \code{\link[=hdMTD_CUT]{hdMTD_CUT()}} documentation.
\item For "BIC" method, extra parameters are listed in \code{\link[=hdMTD_BIC]{hdMTD_BIC()}} documentation.
}
}
\examples{
X <- perfectSample(MTDmodel(c(1,5),c(0,1)),5000)
hdMTD(X=X,d=5, method = "FS",l=2)
hdMTD(X=X,d=5, method = "FSC",alpha=0.001,xi=1,l=3)
hdMTD(X=X,d=5, method = "BIC",xi=1, minl=3, maxl=3)
hdMTD(X=X,d=5, method = "CUT",S=c(1,5,3,2),alpha=0.1,xi=1)

}
