% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdMTD_CUT.R
\name{hdMTD_CUT}
\alias{hdMTD_CUT}
\title{The Cut method for inference in MTD models}
\usage{
hdMTD_CUT(
  X,
  d,
  S = 1:d,
  alpha = 0.05,
  mu = 1,
  xi = 0.5,
  A = NULL,
  warning = FALSE,
  ...
)
}
\arguments{
\item{X}{A vector or single-column data frame containing a chain sample.}

\item{d}{A positive integer representing an upper bound for the chain order.}

\item{S}{A numeric vector of positive integers from which this function will select
a set of relevant lags. Typically, \code{S} is a subset of \code{1:d}. If \code{S}
is not provided, by default \code{S=1:d}.}

\item{alpha}{A positive real number, \code{alpha}, appears in a threshold used in the CUT
step to determine if two distributions are different enough. The larger the \code{alpha},
the greater the distance required to consider that there is a difference between a set
of distributions.}

\item{mu}{A positive real number between 0 and 3. \code{mu} is also a component of the same
threshold as \code{alpha}.}

\item{xi}{A positive real number, \code{xi} is also a component of the same threshold as
\code{alpha} and \code{mu}.}

\item{A}{A vector with positive integers representing the state space. If not informed,
this function will set \code{A=unique(X)}.}

\item{warning}{Logical. If \code{TRUE}, informs the user if the state space was set as
\code{A=unique(X)}.}

\item{...}{Other parameters. This is used to accommodate any additional argument passed
to \code{\link[=hdMTD_CUT]{hdMTD_CUT()}} through the \code{\link[=hdMTD]{hdMTD()}} function.}
}
\value{
Returns a set of relevant lags estimated using the CUT algorithm.
}
\description{
A function that estimates the set of relevant lags of an MTD model using the CUT method.
}
\details{
The "Forward Stepwise and Cut" (FSC) is an algorithm for inference in
Mixture Transition Distribution (MTD) models. It consists
in the application of the "Forward Stepwise" (FS) step followed by the CUT algorithm.
This method and its steps where developed by \href{http://jmlr.org/papers/v24/22-0266.html}{Ost and Takahashi}
and are specially useful for inference in high-order MTD Markov chains. This specific function
will only apply the CUT step of the algorithm and return an estimated relevant lag set.
}
\examples{
X <- testChains[,3]
hdMTD_CUT(X,4,alpha=0.02,mu=1,xi=0.4)
hdMTD_CUT(X,d=6,S=c(1,4,6),alpha=0.0065)

}
\references{
Ost, G. and Takahashi, D. Y. (2023), "Sparse markov models for high-dimensional inference",
Journal of Machine Learning Research 24(279), 1-54. \href{http://jmlr.org/papers/v24/22-0266.html}{URL: http://jmlr.org/papers/v24/22-0266.html}
}
