% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdMTD_FS.R
\name{hdMTD_FS}
\alias{hdMTD_FS}
\title{The Foward Stepwise method.}
\usage{
hdMTD_FS(X, d, l, A = NULL, elbowTest = FALSE, warning = FALSE, ...)
}
\arguments{
\item{X}{A mixture transition distribution (MTD) chain sample.}

\item{d}{An upper bound for the chains order.}

\item{l}{Stop point for FS algorithm.}

\item{A}{The States space.}

\item{elbowTest}{If TRUE the function will choose the point where the \eqn{ max(\nu)} stops decreasing (that is, where the graph of the \eqn{ max(\nu)} changes its direction like an elbow).
If l is smaller than this point the function will simply return a set with l elements, so setting elbowTest=TRUE only makes a difference if the elbow point happens in the first l-1 elements of \eqn{ max(\nu)}.}

\item{warning}{If True may return warnings.}

\item{...}{Used to accommodate any extra arguments passed by the \code{\link[=hdMTD]{hdMTD()}} function.}
}
\value{
Returns a estimated S set of relevant lags with size l .
}
\description{
A function for inference in MTD Markov chains with FS method. Applies Forward Stepwise (FS) algorithm to estimate a relevant lag set \eqn{\Lambda} for MTD models.
}
\details{
The "Forward Stepwise" (FS) algorithm is the first step of the "Forward Stepwise and Cut" (FSC) algorithm for inference in
Mixture Transition Distribution (MTD) models. Which consists in the application of the "Forward Stepwise" (FS) step followed by the CUT algorithm.
This method and its steps where developed by \href{https://arxiv.org/abs/2202.08007}{Ost and Takahashi} and are specially useful for inference in High order MTD Markov chains.
This specific function will apply only the FS step of the algorithm and return an estimated relevant lag set of size l.
}
\examples{
X <- perfectSample(MTDmodel(Lambda=c(2,4),A=c(0,1),w0=0.05),2000)
hdMTD_FS(X,A=c(0,1),d=5,l=2)
hdMTD_FS(X,d=5,l=2)
hdMTD_FS(X,d=4,l=3,elbowTest = TRUE)
hdMTD_FS(X,d=4,l=3,elbowTest = TRUE)

}
