% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdMTD_FS.R
\name{hdMTD_FS}
\alias{hdMTD_FS}
\title{The Forward Stepwise (FS) method for inference in MTD models}
\usage{
hdMTD_FS(X, d, l, A = NULL, elbowTest = FALSE, warning = FALSE, ...)
}
\arguments{
\item{X}{A vector or single-column data frame containing a chain sample.}

\item{d}{A positive integer representing an upper bound for the chain order.}

\item{l}{A positive integer that sets the number of elements in the output vector.}

\item{A}{A vector with positive integers representing the state space. If not informed,
this function will set \code{A=unique(X)}.}

\item{elbowTest}{Logical. If TRUE, the function will use a special criterion to determine the length
of the estimated set of relevant lags. See \emph{Details} below for more information.}

\item{warning}{Logical. If \code{TRUE}, informs the user if the state space was set as \code{A=unique(X)}.}

\item{...}{Other parameters. This is used to accommodate any additional argument passed
to \code{\link[=hdMTD_FS]{hdMTD_FS()}} through the \code{\link[=hdMTD]{hdMTD()}} function.}
}
\value{
Returns a vector with the estimated relevant lag set using FS algorithm.
}
\description{
A function that estimates the set of relevant lags of an MTD model using the FS method.
}
\details{
The "Forward Stepwise" (FS) algorithm is the first step of the "Forward Stepwise and Cut"
(FSC) algorithm for inference in Mixture Transition Distribution (MTD) models.
This method was developed by \href{https://arxiv.org/abs/2202.08007}{Ost and Takahashi}
This specific function will only apply the FS step of the algorithm and return an estimated
relevant lag set of length \code{l}.

If the algorithm determines that there are multiple lags equally important and more
important than all others, it will sample one of them uniformly.

If \code{elbowTest=TRUE}, the function will use a new criterion to determine the length of
the estimated relevant lag set. In the FS algorithm, a certain quantity \eqn{\nu} is calculated
for each lag in \code{1:d}, and the lag with the greatest \eqn{\nu} is deemed important. This lag
is included in the output, and using this knowledge, the function proceeds to seek the next
important lag (the one with the highest \eqn{\nu} among the remaining ones). The process stops
when the output vector reaches length \code{l} if \code{elbowTest=FALSE}. If \code{elbowTest=TRUE},
the function will store a vector with the value of these greatest \eqn{\nu} at each step. It will
then look at this vector and identify the position of the smallest \eqn{\nu} among them. The
output will only keep the lags that came before the one responsible for this \eqn{\nu} value.
}
\examples{
X <- testChains[,1]
hdMTD_FS(X,d=5,l=2)
hdMTD_FS(X,d=4,l=3,elbowTest = TRUE)

}
\references{
Ost, G. and Takahashi, D. (2022), ‘Sparse markov models for high-dimensional inference’.
\href{https://arxiv.org/abs/2202.08007}{ arXiv:2202.08007}
}
