% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdMTD_FSC.R
\name{hdMTD_FSC}
\alias{hdMTD_FSC}
\title{Forward Stepwise and Cut method for inference in MTD models}
\usage{
hdMTD_FSC(
  X,
  d,
  l,
  alpha = 0.05,
  mu = 1,
  xi = 0.5,
  A = NULL,
  warning = FALSE,
  ...
)
}
\arguments{
\item{X}{A vector or single-column data frame containing a chain sample.}

\item{d}{A positive integer representing an upper bound for the chain order.}

\item{l}{A positive integer that sets the number of elements in the output vector.}

\item{alpha}{A positive real number, \code{alpha}, appears in a threshold used in the CUT
step to determine if two distributions are different enough. The larger the \code{alpha},
he greater the distance required to consider that there is a difference between a set
of distributions.}

\item{mu}{A positive real number between 0 and 3. \code{mu}is also a component of the same
threshold as \code{alpha}.}

\item{xi}{A positive real number, \code{xi} is also a component of the same threshold as
\code{alpha} and \code{mu}.}

\item{A}{A vector with positive integers representing the state space. If not informed,
this function will set \code{A=unique(X)}.}

\item{warning}{Logical. If \code{TRUE}, informs the user if the state space was set as
\code{A=unique(X)}.}

\item{...}{Other parameters. This is used to accommodate any additional argument passed
to \code{\link[=hdMTD_FSC]{hdMTD_FSC()}} through the \code{\link[=hdMTD]{hdMTD()}} function.}
}
\value{
Returns a vector with the estimated relevant lag set using FSC algorithm.
}
\description{
A function for inference in MTD Markov chains with FSC method. This function estimates the relevant
lag set \eqn{\Lambda} of a MTD model through the FSC algorithm.
}
\details{
The "Forward Stepwise and Cut" (FSC) is an algorithm for inference in
Mixture Transition Distribution (MTD) models. It consists
in the application of the "Forward Stepwise" (FS) step followed by the CUT algorithm.
This method and its steps where developed by \href{https://arxiv.org/abs/2202.08007}{Ost and Takahashi}
and are specially useful for inference in high-order MTD Markov chains.

#'# Author(s):
This method was developed in \href{https://arxiv.org/abs/2202.08007}{Ost and Takahashi},
(2022), "Sparse markov models for high-dimensional inference".
}
\examples{
X <- testChains[,1]
hdMTD_FSC(X,4,3,alpha=0.02)
hdMTD_FSC(X,4,2,alpha=0.001)

}
