% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTDest.R
\name{MTDest}
\alias{MTDest}
\title{EM estimation}
\usage{
MTDest(X, S, M = 1, init, A = NULL)
}
\arguments{
\item{X}{A MTD chain sample}

\item{S}{The relevant lag set}

\item{M}{A stopping point for the EM algorithm.}

\item{init}{A list with initial parameters.}

\item{A}{The states space}
}
\value{
A list with the estimated parameters of the MTD model
}
\description{
Estimation of MTD parameters through EM algorithm
}
\details{
About the parameter M: it serves as a stopping
criterion for the EM algorithm. If the distance between
the log-likelihood calculated with the newly estimated
parameters and the log-likelihood calculated with the
previous parameters is less than M, the algorithm stops.

About the parameter init: It's a list that must have 3
entries: a vector named p0 (representing an independent
distribution, ordered from the smallest to the greatest
element of A), p_j (a list containing a stochastic matrix
for each element of S, also ordered from the smallest to
the greatest element of S), and a vector named lambdas
(representing the weights, first the weight for p0, and
then for each element in p_j).
}
