% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTDest.R
\name{MTDest}
\alias{MTDest}
\title{Title}
\usage{
MTDest(X, S, M = 2, init, A = NULL, init_ask = FALSE, indep_part = TRUE)
}
\arguments{
\item{X}{bla}

\item{S}{bla}

\item{M}{bla}

\item{init}{bla}

\item{A}{bla}

\item{init_ask}{bla}

\item{indep_part}{bla}
}
\value{
bla
}
\description{
Title
}
