% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTDmodel.R
\name{MTDmodel}
\alias{MTDmodel}
\title{Creates an MTD model}
\usage{
MTDmodel(
  Lambda,
  A,
  lam0 = NULL,
  lamj = NULL,
  pj = NULL,
  p0 = NULL,
  single_matrix = FALSE,
  indep_part = TRUE
)
}
\arguments{
\item{Lambda}{The relevant lag set should consist of positive integers and will be sorted from smallest
to greatest. The smallest number represents the latest (most recent) time in the past, and the greatest
number represents the earliest time in the past.}

\item{A}{The states space.}

\item{lam0}{The weight of the independent distribution, must be a value in \verb{[0,1)}.}

\item{lamj}{A vector of weights for the distributions pj. Values must be in the range \verb{[0, 1)}. The order of weights must match the order
of the sorted Lambda set.}

\item{pj}{A list with lenL matrices of size \verb{lenA x lenA}. Here, lenL represents the number of elements in Lambda, and lenA represents the number
of elements in A.}

\item{p0}{A vector with the independent distribution part of an MTD model. If not informed and argument indep_part=TRUE, the distribution will be sampled
from a uniform distribution. If indep_part=FALSE, then there is no independent distribution and p0 entries will be set to zero. If you enter p0=0,
then indep_part will be set to FALSE.}

\item{single_matrix}{If TRUE all pj matrices are equal.}

\item{indep_part}{If FALSE there is no independent distribution and p0 is set to zero.}
}
\value{
An MTD model as a MTD object.
}
\description{
Given a set of parameters generates a Mixture transition distribution model as an object of class MTD.
}
\details{
This MTD object can be used by functions such as \code{\link[=oscillation]{oscillation()}}, which retrieves the model's oscillation, and \code{\link[=perfectSample]{perfectSample()}},
which will perfectly sample a MTD Markov chain with the model's parameters
}
\examples{
MTDmodel(Lambda=c(1,3),A=c(4,8,12))

MTDmodel(Lambda=c(2,4,9),A=c(0,1),lam0=0.05,lamj=c(0.35,0.2,0.4),
pj=list(matrix(c(0.5,0.7,0.5,0.3),ncol=2)),p0=c(0.2,0.8),single_matrix=TRUE)

MTDmodel(Lambda=c(2,4,9),A=c(0,1),lam0=0.05,
pj=list(matrix(c(0.5,0.7,0.5,0.3),ncol=2)),single_matrix=TRUE,indep_part=FALSE)
}
