% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTDmodel.R
\name{MTDmodel}
\alias{MTDmodel}
\title{Creates an MTD model}
\usage{
MTDmodel(
  Lambda,
  A,
  w0 = NULL,
  w_j = NULL,
  p_j = NULL,
  p0 = NULL,
  single_matrix = FALSE,
  indep_part = TRUE
)
}
\arguments{
\item{Lambda}{The relevant lag set. Should be positive integers and will be sorted from smallest to greater. The smallest number
will represent the latest \verb{(most recent)} time in the past and the greatest number the earliest time in the past.}

\item{A}{The states space.}

\item{w0}{The weight of the independent distribution, must be a value in \verb{[0,1)}.}

\item{w_j}{A vector of weights for the distributions p_j. Values must be in \verb{[0,1)}. The order of weights must match the order
of the sorted Lambda set.}

\item{p_j}{A list with lenL matrices \verb{lenA x lenA}. Where lenL is the number of elements in Lambda and lenA the number of elements in A.}

\item{p0}{A vector with the independent distribution part of a MTD model. If not informed and indep_part=TRUE the distribution will be sampled
form a uniform. If indep_part=FALSE the distribution will be set to zero.}

\item{single_matrix}{If TRUE all p_j matrix are equal.}

\item{indep_part}{If FALSE independent distribution is set to zero.}
}
\value{
An MTD model as a MTD object.
}
\description{
Given a set of parameters generates a Mixture transition distribution model as an object of class MTD.
}
\details{
This MTD object can be used by functions such as \code{\link[=oscillation]{oscillation()}} which retrieves the models oscillation and \code{\link[=perfectSample]{perfectSample()}} which will
perfectly sample a Markov chain with the models parameters.
}
\examples{
MTDmodel(Lambda=c(1,3),A=c(4,8,12))

MTDmodel(Lambda=c(2,4,9),A=c(0,1),w0=0.05,w_j=c(0.35,0.2,0.4),
p_j=list(matrix(c(0.5,0.7,0.5,0.3),ncol=2)),p0=c(0.2,0.8),single_matrix=TRUE)

MTDmodel(Lambda=c(2,4,9),A=c(0,1),w0=0.05,
p_j=list(matrix(c(0.5,0.7,0.5,0.3),ncol=2)),single_matrix=TRUE,indep_part=FALSE)
}
