% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTDmodel.R
\name{MTDmodel}
\alias{MTDmodel}
\title{Creates an MTD model}
\usage{
MTDmodel(
  Lambda,
  A,
  lam0 = NULL,
  lamj = NULL,
  pj = NULL,
  p0 = NULL,
  single_matrix = FALSE,
  indep_part = TRUE
)
}
\arguments{
\item{Lambda}{A numeric vector of positive integers representing the relevant lag set.
The elements will be sorted from smallest to greatest. The smallest number represents the latest
(most recent) time in the past, and the greatest number represents the earliest time in the past.}

\item{A}{A vector with positive integers representing the state space.}

\item{lam0}{The weight of the independent distribution, must be a value in \verb{[0,1)}.}

\item{lamj}{A vector of weights for the distributions in the argument \code{pj}. Values must be
in the range \verb{[0, 1)}. The first element in \code{lamj} must be the weight of the first
element of the list \code{pj} and so on.}

\item{pj}{A list with \code{length(Lambda)} stochastic matrices.}

\item{p0}{A vector with the independent distribution part of an MTD model. If not informed
and argument \code{indep_part=TRUE}, the distribution will be sampled from a uniform distribution.
If \code{indep_part=FALSE}, then there is no independent distribution and \code{p0} entries will
be set to zero. If you enter \code{p0=0}, then \code{indep_part} will be set to \code{FALSE}.}

\item{single_matrix}{Logical. If \code{TRUE}, all matrices in list \code{pj} are equal.}

\item{indep_part}{Logical. If \code{FALSE} there is no independent distribution and \code{p0}
is set to zero.}
}
\value{
An MTD model as an MTD object.
}
\description{
Given a set of parameters it generates an MTD model as an object of class MTD.
}
\details{
This MTD object can be used by functions such as \code{\link[=oscillation]{oscillation()}}, which retrieves the
model's oscillation, and \code{\link[=perfectSample]{perfectSample()}}, which will perfectly sample an MTD Markov chain
with the model's parameters.
}
\examples{
MTDmodel(Lambda=c(1,3),A=c(4,8,12))

MTDmodel(Lambda=c(2,4,9),A=c(0,1),lam0=0.05,lamj=c(0.35,0.2,0.4),
pj=list(matrix(c(0.5,0.7,0.5,0.3),ncol=2)),p0=c(0.2,0.8),single_matrix=TRUE)

MTDmodel(Lambda=c(2,4,9),A=c(0,1),lam0=0.05,
pj=list(matrix(c(0.5,0.7,0.5,0.3),ncol=2)),single_matrix=TRUE,indep_part=FALSE)
}
