% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_parameters.R
\name{n_parameters}
\alias{n_parameters}
\title{Number of Parameters in an MTD Model}
\usage{
n_parameters(Lambda, A, single_matrix = FALSE, indep_part = TRUE)
}
\arguments{
\item{Lambda}{A numeric vector of positive integers representing the relevant lag set.
The elements will be sorted from smallest to greatest. The smallest number represents the latest
(most recent) time in the past, and the greatest number represents the earliest time in the past.}

\item{A}{A vector with positive integers representing the state space.}

\item{single_matrix}{Logical. If \code{TRUE}, the model depends on any past in \code{Lambda}
through the same stochastic matrix.}

\item{indep_part}{Logical. If \code{FALSE} there is no independent distribution.}
}
\value{
The number of parameters in an MTD model.
}
\description{
Calculates the number of parameters in an MTD model with a relevant lag set Lambda and state space A.
}
