% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oscillation.R
\name{oscillation}
\alias{oscillation}
\title{Calculates the oscillation of a MTD model}
\usage{
oscillation(MTD)
}
\arguments{
\item{MTD}{an MTD object.}
}
\value{
Returns the oscillations of a MTD object, that is, for each
relevant lag j, returns lambda_j times the maximum of the total variation
distance between the distributions in matrix p_j.
}
\description{
Calculates the oscillation of a MTD model
}
