% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oscillation.R
\name{oscillation}
\alias{oscillation}
\title{Oscillations of a MTD model}
\usage{
oscillation(x)
}
\arguments{
\item{x}{Must be a MTD object.}
}
\value{
Returns the oscillations for each relevant lag of a MTD object.
}
\description{
Calculates the oscillations of a MTD model object.
}
\details{
The oscillations of a MTD model
(\eqn{\delta_k} for \eqn{k in \Lambda}), are the product of the weight \eqn{\lambda_k}
multiplied by the maximum of the total variation distances between the
distributions in the matrix p_k. These values are important because they
measure the influence of a relevant lag k on the model.
}
\examples{
oscillation( MTDmodel(Lambda=c(1,4),A=c(2,3) ) )
oscillation( MTDmodel(Lambda=c(1,4),A=c(2,3),w0=0.01 ) )
oscillation( MTDmodel(Lambda=c(1,4),A=c(2,3),w0=0.01,w_j=c(0.49,0.5)) )
oscillation( MTDmodel(Lambda=c(1,4),A=c(2,3),w0=0.01,w_j=c(0.49,0.5),
p_j=list(matrix(c(0.1,0.9,0.9,0.1),ncol=2)), single_matrix=TRUE))

}
