% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oscillation.R
\name{oscillation}
\alias{oscillation}
\title{Oscillations of a MTD Markov chain}
\usage{
oscillation(x, ...)
}
\arguments{
\item{x}{Must be a MTD object or a chain sample.}

\item{...}{Additional parameters that might be required. Such as:

S: If x is a chain sample,
the user should provide a set of lags, that must be labeled as S, for which he wishes to estimate the oscillations.
In this case d=max{S}.

A: If x is a chain sample, and there may be elements is A that did not appear in x, the state space A should be
specified, and it must be labeled as A.}
}
\value{
If the x parameter is an MTD object, it will provide the oscillations for
each relevant element. In case x is a MTD chain sample, it estimates the oscillations
for a user-inputted set S of lags.
}
\description{
Calculates the oscillations of a MTD model object or estimates the oscillations of a chain sample.
}
\details{
The oscillations of a MTD model
(\eqn{\delta_j} for \eqn{j in \Lambda}), are the product of the weight \eqn{\lambda_j}
multiplied by the maximum of the total variation distances between the
distributions in the matrix pj. These values are important because they
measure the influence of a relevant lag j on the model.
}
\examples{
oscillation( MTDmodel(Lambda=c(1,4),A=c(2,3) ) )
oscillation( MTDmodel(Lambda=c(1,4),A=c(2,3),lam0=0.01 ) )
oscillation( MTDmodel(Lambda=c(1,4),A=c(2,3),lam0=0.01,lamj=c(0.49,0.5)) )
oscillation( MTDmodel(Lambda=c(1,4),A=c(2,3),lam0=0.01,lamj=c(0.49,0.5),
pj=list(matrix(c(0.1,0.9,0.9,0.1),ncol=2)), single_matrix=TRUE))

}
