% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oscillation.R
\name{oscillation}
\alias{oscillation}
\title{Oscillations of a MTD model}
\usage{
oscillation(MTD)
}
\arguments{
\item{MTD}{an MTD object.}
}
\value{
Returns the oscillations for each relevant lag of a MTD object.
}
\description{
Calculates the oscillations of a MTD model object.
}
\details{
The oscillations of a MTD model
(\eqn{\delta_j} for \eqn{j in \Lambda}), are the product of the weight \eqn{\lambda_j} times the maximum of the total variation distances between the distributions in the matrix p_j.
These values are important because they measure how much a relevant lag j influences the model.
}
