% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfectSample.R
\name{perfectSample}
\alias{perfectSample}
\title{Perfect Sample.}
\usage{
perfectSample(MTD, N)
}
\arguments{
\item{MTD}{a canonical MTD object.}

\item{N}{the sample size. If NULL sample size will be set to 1000.}
}
\value{
Returns a sample from o MTD model.
}
\description{
Perfectly samples a MTD Markov chain.
}
\details{
Can sample a MTD Markov Chain with stationary distribution.
}
