% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfectSample.R
\name{perfectSample}
\alias{perfectSample}
\title{Perfect Sample.}
\usage{
perfectSample(MTD, N = NULL)
}
\arguments{
\item{MTD}{An MTD object. The state space A in the MTD object must be sorted}

\item{N}{The sample size. If NULL sample size will be set to 1000.}
}
\value{
Returns a sample from o MTD model (the first element is the most recent).
}
\description{
Perfectly samples a MTD Markov chain.
}
\details{
Samples a MTD Markov Chain from stationary distribution.

This perfect sample algorithm requires that the MTD model has
an independent distribution and a positive weight for it.
}
\examples{
perfectSample(MTDmodel(Lambda=c(1,4),A=c(0,2)),N=200 )
perfectSample(MTDmodel(Lambda=c(2,5),A=c(1,2,3)),N=1000 )

}
