% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PI.R
\name{PI}
\alias{PI}
\title{Estimated stationary distribution vector for every a_j and a sequence x_S}
\usage{
PI(S, base, x_S, lenX, d)
}
\arguments{
\item{S}{Set of lags.}

\item{base}{A table with frequencies of a Markov Chain sample.}

\item{x_S}{A specific sequence indexed by S.}

\item{lenX}{Lenght of sample vector.}

\item{d}{The order.}
}
\value{
Returns a vector of estimated stationary distributions of a sequence X_S
with the element in j varying in all the states space.
}
\description{
Estimated stationary distribution vector for every a_j and a sequence x_S
}
\details{
#(pi(xa_Sj),pi(xb_Sj),pi(xc_Sj),...)
}
