% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probs.R
\name{probs}
\alias{probs}
\title{Estimated transition probabilities}
\usage{
probs(X, S, matrixform = FALSE, A = NULL, warning = FALSE)
}
\arguments{
\item{X}{A vector or single-column data frame containing a chain sample.}

\item{S}{A numeric vector of positive integers representing a set of relevant lags.}

\item{matrixform}{Logical. If \code{TRUE}, the output is formatted as a stochastic transition matrix.}

\item{A}{A numeric vector representing the state space, consisting of distinct integers.
If not provided, this function will set \code{A = unique(X)}.}

\item{warning}{Logical. If \code{TRUE}, the function warns the user when the state space is automatically set as \code{A = unique(X)}.}
}
\value{
The MLE estimates for the given chain sample, conditioned on the relevant lags in \code{S}.
If \code{matrixform = TRUE}, the result is a stochastic matrix.
}
\description{
Computes the Maximum likelihood estimators (MLE) for an MTD Markov chain with
relevant lag set \code{S}.
}
\examples{
X <- testChains[, 3]
probs(X, S = c(1, 30))
probs(X, S = c(1, 15, 30))

}
