% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probs.R
\name{probs}
\alias{probs}
\title{Estimated transition probabilities}
\usage{
probs(X, S, A = NULL, warning = FALSE)
}
\arguments{
\item{X}{Markov Chain.}

\item{S}{Relevant lag set.}

\item{A}{State space.}

\item{warning}{If TRUE may return warnings.}
}
\value{
The MLE for a given Markov Chain sample with relevant lag set S.
}
\description{
Maximum likelihood estimators for a Markov Chain with relevant lag set S.
}
\examples{
X <- perfectSample(MTDmodel(Lambda=c(1,7),A=c(1,2)),N=500)
probs(X,S=c(1,7))
probs(X,S=c(1,7,9))

}
