% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probs.R
\name{probs}
\alias{probs}
\title{Estimated transition probabilities}
\usage{
probs(X, S, matrixform = FALSE, A = NULL, warning = FALSE)
}
\arguments{
\item{X}{A MTD Markov Chain.}

\item{S}{The estimated relevant lags set.}

\item{matrixform}{If TRUE, the function returns probability estimates in the form of a stochastic matrix.}

\item{A}{The state space. "A" only needs to be informed if X does not already contain all elements of "A"}

\item{warning}{If TRUE may return warnings.}
}
\value{
The MLE for a given MTD Markov Chain sample with relevant lags set S.
}
\description{
Maximum likelihood estimators for a MTD Markov chain with relevant lag set S.
}
\examples{
X <- perfectSample(MTDmodel(Lambda=c(1,7),A=c(1,2)),N=500)
probs(X,S=c(1,7))
probs(X,S=c(1,7,9))

}
