% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeSample.R
\name{shapeSample}
\alias{shapeSample}
\title{Creates a table with sample information}
\usage{
shapeSample(X, d)
}
\arguments{
\item{X}{A Markov chain sample.}

\item{d}{The order of the Markov chain.}
}
\value{
A table with every sequence in the sample and its frequency.
}
\description{
Creates a table with sample information
}
\details{
The function will make a tibble with \eqn{d+2} columns. For each row, the first \eqn{d+1} columns will
have a sequency of size \eqn{d+1} that appeared in the sample. The last column, called \code{Nxa},
will contain the number of times each of these sequences appeared in the sample.
}
\examples{
shapeSample(c(1,2,2,1,2,1,1,2,1,2),3)
shapeSample(c(0,2,0,2,0,2,1,1,0,0,1,2,1,2,1),4)
}
