% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseMarkov.R
\name{sparseMarkov}
\alias{sparseMarkov}
\title{A function for inference in sparse Markov chains}
\usage{
sparseMarkov(X, d, method, l = NULL)
}
\arguments{
\item{X}{A markov chain sample}

\item{d}{The order or a upper bound for the order.}

\item{method}{A method for estimation}

\item{l}{A stop point for some methods}
}
\value{
The relevant lag set
}
\description{
A function for inference in sparse Markov chains
}
