% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseMarkov_BIC.R
\name{sparseMarkov_BIC}
\alias{sparseMarkov_BIC}
\title{Estimates a relevant lag set Lambda with BIC}
\usage{
sparseMarkov_BIC(X, A, d, l, c = 1/2)
}
\arguments{
\item{X}{A Markov chain.}

\item{A}{The state space.}

\item{d}{A upper threshold for the chains order.}

\item{l}{A upper threshold for the number of elements in the relevant lag set.}

\item{c}{The BIC constant. Defaulted to 1/2. Smaller c \verb{(near 0)} reduces the impact of overparameterization.}
}
\value{
Return a estimator of the relevant lag set using BIC.
}
\description{
Estimates a relevant lag set Lambda with BIC
}
