% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseMarkov_CUT.R
\name{sparseMarkov_CUT}
\alias{sparseMarkov_CUT}
\title{Applies Cut algorithm}
\usage{
sparseMarkov_CUT(X, A, d, alpha, mu, xi, S = 1:d)
}
\arguments{
\item{X}{A mixture transition distribution chain sample.}

\item{A}{The states space.}

\item{d}{The chains order.}

\item{alpha}{A parameter of CUT.}

\item{mu}{A parameter of CUT.}

\item{xi}{A parameter of CUT.}

\item{S}{A set of relevant lags.}
}
\value{
Returns a set that contains relevant lags.
}
\description{
Applies Cut algorithm
}
