% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseMarkov_CUT.R
\name{sparseMarkov_CUT}
\alias{sparseMarkov_CUT}
\title{Aplies Cut algorithm}
\usage{
sparseMarkov_CUT(X, A, d, alpha, mu, xi, S = 1:d)
}
\arguments{
\item{X}{A Markov Chain sample with mixture transition distribution.}

\item{A}{States space.}

\item{d}{Chains order.}

\item{alpha}{A parameter of CUT}

\item{mu}{A parameter of CUT}

\item{xi}{A parameter of CUT}

\item{S}{A set of relevant lags}
}
\value{
Returns a setthat contains relevant lags.
}
\description{
Aplies Cut algorithm
}
