% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseMarkov_FS.R
\name{sparseMarkov_FS}
\alias{sparseMarkov_FS}
\title{Aplies Forward Stepwise (FS) algorithm}
\usage{
sparseMarkov_FS(X, l, A, d)
}
\arguments{
\item{X}{A Markov Chain sample with mixture transition distribution.}

\item{l}{Stop point for FS algorithm.}

\item{A}{States space.}

\item{d}{Chains order.}
}
\value{
Returns a set S of size l that contains relevant lags.
}
\description{
Aplies Forward Stepwise (FS) algorithm
}
